fix: circular dependency between ClassroomService and AttendanceService by @Lazy, and make Dependency Inversion by creating service interfaces
